#10 more tests for jboss_fullconfig
