+++ lib/puppet_x/coi/jboss/configuration.rb
@@ -7,8 +7,13 @@ module Jboss
 # A class for JBoss configuration
 class Configuration
   class << self
-  
+
     @config = nil
+
+    # Test method that return current version(for comatability with ruby 1.8)
+    def ruby_version
+      RUBY_VERSION
+    end
 
     # Gets the main config file
     def configfile
@@ -65,7 +70,7 @@ class Configuration
       @config = value
       nil
     end
-    
+
     # Gets configuration value by its symbol
     #
     # @param key [Symbol] a key in hash
