+++ spec/unit/facter/jboss_fullconfig_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 require 'puppet_x/coi/jboss/configuration'
 
 describe 'Fact jboss_fullconfig', :type => :fact do
+
   let(:sample_config) do
     t = Tempfile.new('rspec-jboss-fullconfig')
     path = t.path
@@ -13,44 +14,44 @@ describe 'Fact jboss_fullconfig', :type => :fact do
     # The Jboss home directory.
     #
     JBOSS_HOME=/usr/lib/wildfly-12.2.0.Final
-    
+
     # The JBoss product name.
     #
     JBOSS_PRODUCT=wildfly
-    
+
     # The JBoss version.
     #
     JBOSS_VERSION=12.2.0.Final
-    
+
     # The JBoss configuration file
     #
     JBOSS_CONFIG=standalone-full.xml
-    
+
     # The username who should own the process.
     #
     JBOSS_USER=wildfly
-    
+
     # The amount of time to wait for startup
     #
     # STARTUP_WAIT=30
-    
+
     # The amount of time to wait for shutdown
     #
     # SHUTDOWN_WAIT=30
-    
+
     # Location to keep the console log
     #
     JBOSS_CONSOLE_LOG=/var/log/wildfly/console.log
-    
+
     # Runs JBoss in domain mode?
     JBOSS_RUNASDOMAIN=false
-    
+
     # JBoss running mode: domain or standalone
     JBOSS_MODE=standalone
-    
+
     # Default JBoss domain controller
     JBOSS_CONTROLLER=127.0.0.1:9990
-    
+
     # Default JBoss domain profile
     JBOSS_PROFILE=full
     eos
@@ -89,13 +90,25 @@ describe 'Fact jboss_fullconfig', :type => :fact do
     File.unlink(sample_config)
   end
   subject { Facter.value(:jboss_fullconfig) }
+  shared_examples 'is not nill and empty' do
+    it { expect(subject).not_to be_nil }
+    it { expect(subject).not_to be_empty }
+  end
   context 'with sample config file for WildFly 8.2' do
-    context 'return value' do
-      it { expect(subject).not_to be_nil }
-      it { expect(subject).not_to be_empty }
+    context 'without mocking RUBY_VERSION' do
+      it_behaves_like 'is not nill and empty'
       it { expect(subject).to respond_to(:[]) }
       its(:size) { should eq(10) }
       it { expect(subject).to eq(expected_hash) }
+    end
+
+    context 'with mocking RUBY_VERSION to 1.8.7' do
+      before :each do
+        expect(Puppet_X::Coi::Jboss::Configuration).to receive(:ruby_version).once.and_return('1.8.7')
+      end
+
+      # it_behaves_like 'is not nill and empty'
+      it { expect(subject.to_s).to eq("{\"home\"=>\"/usr/lib/wildfly-12.2.0.Final\", \"product\"=>\"wildfly\", \"version\"=>\"12.2.0.Final\", \"config\"=>\"standalone-full.xml\", \"user\"=>\"wildfly\", \"console_log\"=>\"/var/log/wildfly/console.log\", \"runasdomain\"=>false, \"mode\"=>\"standalone\", \"controller\"=>\"127.0.0.1:9990\", \"profile\"=>\"full\"}")}
     end
   end
 end
